Set expiry to None for test cache

+++ Simyan/sqlite_cache.py
@@ -5,7 +5,7 @@ from typing import Any, Dict, Optional
 
 
 class SqliteCache:
-    def __init__(self, name: str = "Simyan-Cache.sqlite", expiry: int = 14):
+    def __init__(self, name: str = "Simyan-Cache.sqlite", expiry: Optional[int] = 14):
         self.expiry = expiry
         self.con = sqlite3.connect(name)
         self.cur = self.con.cursor()
@@ -13,10 +13,16 @@ class SqliteCache:
         self.delete()
 
     def select(self, query: str) -> Dict[str, Any]:
-        self.cur.execute(
-            "SELECT response FROM queries WHERE query = ? AND expiry > ?;",
-            (query, datetime.now().strftime("%Y-%m-%d")),
-        )
+        if self.expiry:
+            self.cur.execute(
+                "SELECT response FROM queries WHERE query = ? AND expiry > ?;",
+                (query, datetime.now().strftime("%Y-%m-%d")),
+            )
+        else:
+            self.cur.execute(
+                "SELECT response FROM queries WHERE query = ?;",
+                (query,),
+            )
         result = self.cur.fetchone()
         if result:
             return json.loads(result[0])
@@ -37,6 +43,8 @@ class SqliteCache:
         return self.insert(query=key, response=value)
 
     def delete(self):
+        if not self.expiry:
+            return
         self.cur.execute(
             "DELETE FROM queries WHERE expiry < ?;",
             (datetime.now().strftime("%Y-%m-%d"),),
